Split Config.jl into 3 files
